fix(rapl): gate pure helpers for non-Linux; split sysfs I/O for coverage
macOS clippy (-D warnings) failed: cpu_share/counter_delta are used only by
the Linux-gated sampler, so they were dead code on non-Linux. Gate them
cfg(any(target_os=linux, test)) so tests still reach them everywhere.

Split the powercap sysfs I/O into src/rapl/sysfs.rs and exclude it from
codecov, mirroring src/perf/syscall.rs: reading energy_uj needs root, so CI
can't exercise it. The pure logic (cpu_share, counter_delta) stays in
src/rapl/mod.rs and is now 100% covered.
